requirements.txt
prometheus-client==0.13.1
speedtest-cli==2.1.3
